chore: rename and move sort filter component
